Splits the Domain Token Information block into a separate class. Improves tests.
